fix match method crop image failed

+++ atx/drivers/mixin.py
@@ -283,7 +283,7 @@ class DeviceMixin(object):
         # image match
         screen = imutils.from_pillow(screen) # convert to opencv image
         if rect and isinstance(rect, tuple) and len(rect) == 4:
-            (x0, y0, x1, y1) = [v*pattern_scale for v in rect]
+            (x0, y0, x1, y1) = [int(v*pattern_scale) for v in rect]
             (dx, dy) = dx+x0, dy+y0
             screen = imutils.crop(screen, x0, y0, x1, y1)
             #cv2.imwrite('cc.png', screen)
